Added plus minus implementation to checkout page and notification real-time change on both drinking and checkout page
