+++ app/views/drinks/_drinks.html.erb
@@ -35,5 +35,4 @@
   </div>
 <% end %>
 
-<%= link_to "Pay", order_path(@order), class: "btn btn-success" %>
 <%= javascript_pack_tag 'drinks' %>
