+++ app/views/drinks/index.js.erb
@@ -1,8 +1,5 @@
 
 
-document.getElementById("drink-list").innerHTML = "<%= j render('/drinks/drinks', drinks: @drinks, order: @order) %>";
-document.getElementById('drink-title').innerText = "<%= j @category.nil? ? 'Popular Drinks' : @category.name %>"
-
 // <%= j render('test.js.erb') %>
 // document.getElementById('eta-btn').click();
 
@@ -12,6 +9,4 @@ document.getElementById('drink-title').innerText = "<%= j @category.nil? ? 'Popu
 // }, 0);
 // let $cartQty = total;
 
-document.querySelector("#shopping-cart-btn .badge").innerText = "<%= @current_order.order_drinks.sum(:amount) %>";
-
-window.initializeCartAnimation();
+<%= render "shared/update_cart" %>
