--- app/views/order_drinks/create.html.erb
@@ -1,2 +1,0 @@
-<h1>OrderDrinks#create</h1>
-<p>Find me in app/views/order_drinks/create.html.erb</p>
