+++ app/views/order_drinks/create.js.erb
@@ -1,0 +1,1 @@
+<%= render "shared/update_cart" %>
