+++ app/views/order_drinks/update.js.erb
@@ -1,0 +1,1 @@
+<%= render "shared/update_cart" %>
