+++ app/views/shared/_update_cart.js.erb
@@ -1,0 +1,5 @@
+document.querySelector("#shopping-cart-btn .badge").innerText = "<%= @current_order.order_drinks.sum(:amount) %>";
+window.initializeCartAnimation();
+
+document.getElementById("drink-list").innerHTML = "<%= j render('/drinks/drinks', drinks: @drinks, order: @current_order) %>";
+document.getElementById('drink-title').innerText = "<%= j @category.nil? ? 'Popular Drinks' : @category.name %>"
