+++ config/routes.rb
@@ -14,6 +14,8 @@ Rails.application.routes.draw do
     end
   end
 
+  resources :order_drinks, only: [:update, :destroy]
+
   get 'orders/:id/confirm', to: 'orders#confirm', as: 'confirm_order'
   get 'orders/:id/review', to: 'orders#review', as: 'review_order'
